backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asn1crypto==1.5.1
boto3==1.36.2
botocore==1.36.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.6
fastapi-cli==0.0.7
filelock==3.16.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
mangum==0.19.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
orjson==3.10.14
packaging==24.2
platformdirs==4.3.6
pluggy==1.5.0
pycparser==2.22
pydantic==2.7.1
pydantic-extra-types==2.10.2
pydantic-settings==2.7.1
pydantic_core==2.16.2
Pygments==2.19.1
PyJWT==2.10.1
pyOpenSSL==24.3.0
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
rich-toolkit==0.13.2
s3transfer==0.11.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
snowflake-connector-python==3.12.4
snowflake-sqlalchemy==1.7.3
sortedcontainers==2.4.0
SQLAlchemy==2.0.37
starlette==0.41.3
tomli==2.2.1
tomlkit==0.13.2
typer==0.15.1
typing_extensions==4.12.2
ujson==5.10.0
urllib3==1.26.20
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==14.1
